requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
pymongo==4.7.3
pydantic==2.7.4
starlette==0.37.2
opencv-python-headless==4.10.0.84
ultralytics==8.3.49
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
scikit-learn==1.3.0
matplotlib==3.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0